SQL_RAG/rag_app/requirements.txt
# SQL RAG with Vertex AI Gemini - Python Dependencies

# Core LangChain packages
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0

# Google Generative AI (Gemini)
google-generativeai>=0.3.0

# Vector store and embeddings
faiss-cpu>=1.7.0
langchain-ollama>=0.1.0  # For embeddings generation (nomic-embed-text)

# Streamlit for web interface
streamlit>=1.28.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Optional dependencies for enhanced features
pyarrow>=10.0.0  # For Parquet file support
rapidfuzz>=3.0.0  # For fuzzy string matching
sqlparse>=0.4.0  # For SQL parsing

# Visualization (optional)
pyvis>=0.3.0  # For network graphs
graphviz>=0.20.0  # For graph visualization

# Development and testing
python-dotenv>=1.0.0  # For environment variable management